backend/requirements.txt
fastapi
uvicorn
langchain
langchain-experimental
langchain-groq
langchain-google-genai
pandas
python-dotenv
tabulate